Add on_delete parameter to all ForeignKey fields
Newer Django version (anything still in support) require a on_delete
parameter to be set on all ForeignKey fields. The value of this field
does not matter here, as we are only using the Django ORM as an
interface to the postfixadmin DB and are using it in readonly mode.